+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## v3.2.0 (2026-04-19)
 
 * Add pre/post processing hooks for custom FileSequence subclasses to control input/output string formats (#154)
+* Create new Usage doc
 
 ## v3.1.1 (2026-03-13)